// src/app/core/enums/codigoTipoCombustivelEnum.ts
export enum CodigoTipoCombustivelEnum {
    ALCOOL = "1",
    GASOLINA = "2",
    DIESEL = "3",
    GASOGENIO = "4",
    GAS_METANO = "5",
    ELETRICO_FONTE_INTERNA = "6",
    ELETRICO_FONTE_EXTERNA = "7",
    GASOLINA_GAS_NATURAL_COMBUSTIVEL = "8",
    ALCOOL_GAS_NATURAL_COMBUSTIVEL = "9",
    DIESEL_GAS_NATURAL_COMBUSTIVEL = "10",
    ALCOOL_GAS_NATURAL_VEICULAR = "12",
    GASOLINA_GAS_NATURAL_VEICULAR = "13",
    DIESEL_GAS_NATURAL_VEICULAR = "14",
    GAS_NATURAL_VEICULAR = "15",
    GASOLINA_ALCOOL = "16",
    GASOLINA_ALCOOL_GAS_NATURAL = "17",
    GASOLINA_ELETRICO = "18",
    GASOLINA_ALCOOL_ELETRICO = "19",
    DIESEL_ELETRICO = "21",
    HIBRIDO = "22",
    HIBRIDO_PLUG_IN = "23",
    ELETRICO = "24"
}

export const CodigoTipoCombustivelDescricao = [
    { value: CodigoTipoCombustivelEnum.ALCOOL, label: 'Álcool' },
    { value: CodigoTipoCombustivelEnum.GASOLINA, label: 'Gasolina' },
    { value: CodigoTipoCombustivelEnum.DIESEL, label: 'Diesel' },
    { value: CodigoTipoCombustivelEnum.GASOGENIO, label: 'Gasogenio' },
    { value: CodigoTipoCombustivelEnum.GAS_METANO, label: 'Gás metano' },
    { value: CodigoTipoCombustivelEnum.ELETRICO_FONTE_INTERNA, label: 'Elétrico fonte interna' },
    { value: CodigoTipoCombustivelEnum.ELETRICO_FONTE_EXTERNA, label: 'Elétrico fonte externa' },
    { value: CodigoTipoCombustivelEnum.GASOLINA_GAS_NATURAL_COMBUSTIVEL, label: 'Gasolina/GNC' },
    { value: CodigoTipoCombustivelEnum.ALCOOL_GAS_NATURAL_COMBUSTIVEL, label: 'Álcool/GNC' },
    { value: CodigoTipoCombustivelEnum.DIESEL_GAS_NATURAL_COMBUSTIVEL, label: 'Diesel/GNC' },
    { value: CodigoTipoCombustivelEnum.ALCOOL_GAS_NATURAL_VEICULAR, label: 'Álcool/GNV' },
    { value: CodigoTipoCombustivelEnum.GASOLINA_GAS_NATURAL_VEICULAR, label: 'Gasolina/GNV' },
    { value: CodigoTipoCombustivelEnum.DIESEL_GAS_NATURAL_VEICULAR, label: 'Diesel/GNV' },
    { value: CodigoTipoCombustivelEnum.GAS_NATURAL_VEICULAR, label: 'Gás natural veicular' },
    { value: CodigoTipoCombustivelEnum.GASOLINA_ALCOOL, label: 'Gasolina/Álcool' },
    { value: CodigoTipoCombustivelEnum.GASOLINA_ALCOOL_GAS_NATURAL, label: 'Gasolina/Álcool/GNV' },
    { value: CodigoTipoCombustivelEnum.GASOLINA_ELETRICO, label: 'Híbrido(Gasolina)' },
    { value: CodigoTipoCombustivelEnum.GASOLINA_ALCOOL_ELETRICO, label: 'Híbrido(Gasolina/Álcool)' },
    { value: CodigoTipoCombustivelEnum.DIESEL_ELETRICO, label: 'Híbrido(Diesel)' },
    { value: CodigoTipoCombustivelEnum.HIBRIDO, label: 'Híbrido' },
    { value: CodigoTipoCombustivelEnum.HIBRIDO_PLUG_IN, label: 'Híbrido(Plug-In' },
    { value: CodigoTipoCombustivelEnum.ELETRICO, label: 'Elétrico' }
];